viz_insight_pipeline/requirements.txt
# Removed streamlit to avoid heavy pyarrow build failures on Python 3.14; not needed for FastAPI service.
# Add lightweight optional visualization libs only if required later.
# streamlit
pandas
numpy
matplotlib
plotly
altair
google-genai
python-dotenv
fastapi
uvicorn
python-multipart
pytest
httpx
# Optional: persist generated profile reports to Cloudinary (recommended for Render where local disk is ephemeral)
cloudinary
# ydata-profiling currently has no compatible release for Python 3.14+.
# Keep it for <=3.13 where profiling reports are supported.
ydata-profiling; python_version < '3.14'
setuptools

# PDF export for profiling reports (works well on Windows)
playwright
